requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.2
altair==5.3.0
antiorm==1.2.1
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.9.0
blis==0.7.11
bs4==0.0.2
cachelib==0.13.0
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudscraper==1.2.71
colorama==0.4.6
comm==0.2.2
comtypes==1.4.7
confection==0.1.5
contourpy==1.2.1
cryptography==43.0.1
cycler==0.12.1
cymem==2.0.8
db==0.1.1
db-sqlite3==0.0.1
debugpy==1.8.2
decorator==5.1.1
Deprecated==1.2.15
distlib==0.3.8
Django==5.1
django-cors-headers==4.4.0
django-ratelimit==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
easyocr==1.7.1
email_validator==2.2.0
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.15.4
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-QRcode==3.2.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.0
flatbuffers==24.3.25
fonttools==4.53.1
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gast==0.6.0
gdown==5.2.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
googlemaps==4.10.0
gps==3.19
gps3==0.33.3
greenlet==3.0.3
grpcio==1.65.5
gunicorn==21.2.0
h11==0.14.0
h5py==3.11.0
huggingface-hub==0.26.2
idna==3.7
imageio==2.34.2
img2pdf==0.5.1
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.26.0
iso8601==2.1.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.5.0
kiwisolver==1.4.5
langcodes==3.4.1
language_data==1.2.0
lazy_loader==0.4
libclang==18.1.1
lxml==5.3.0
Mako==1.3.5
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.4.0
mpmath==1.3.0
msedge-selenium-tools==3.141.4
multidict==6.1.0
murmurhash==1.0.10
mysql-connector-python==9.1.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.5.5.64
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathlib_abc==0.1.1
pathy==0.11.0
pikepdf==9.5.1
pillow==11.1.0
platformdirs==4.2.2
plotly==5.24.0
preshed==3.0.9
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
pure-eval==0.2.2
pyarrow==17.0.0
PyAudio==0.2.14
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==1.10.19
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.2
PyPDF2==3.0.0
pypiwin32==223
PyQRCode==1.2.1
pyserial==3.5
PySocks==1.7.1
pytesseract==0.3.8
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
pyttsx3==2.98
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
qrcode==7.3
referencing==0.35.1
regex==2024.5.15
reportlab==4.2.5
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.19.0
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.2.0
scipy==1.14.0
seaborn==0.13.2
segno==1.6.1
selenium==4.26.1
serial==0.0.97
shapely==2.0.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.4
SQLAlchemy==2.0.32
sqlparse==0.5.1
srsly==2.4.8
stack-data==0.6.3
streamlit==1.36.0
stripe==10.12.0
sympy==1.13.0
tbb==2021.13.0
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
textblob==0.18.0.post0
thinc==8.1.12
threadpoolctl==3.5.0
tifffile==2024.7.2
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.1
torch==2.1.0
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.32.0
trio==0.27.0
trio-websocket==0.11.1
twilio==9.3.1
typer==0.7.0
typing_extensions==4.12.2
tzdata==2024.1
undetected-chromedriver==3.5.5
urllib3==2.2.2
virtualenv==20.26.5
wasabi==1.1.3
watchdog==4.0.1
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
Wikipedia-API==0.7.1
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
XlsxWriter==3.2.0
yarl==1.11.1
gunicorn